requirements.txt
imaplib
email
io
mailparser
python-dotenv
PyPDF2
reportlab